+++ pipeline/tests/powdr_std.rs
@@ -75,13 +75,6 @@ fn split_bb_test() {
 
 #[test]
 #[ignore = "Too slow"]
-fn add_sub16_test() {
-    let f = "std/add_sub16_test.asm";
-    test_plonky3_with_backend_variant::<BabyBearField>(f, vec![], BackendVariant::Composite);
-}
-
-#[test]
-#[ignore = "Too slow"]
 fn arith_test() {
     let f = "std/arith_test.asm";
     let pipeline = make_simple_prepared_pipeline(f);
@@ -117,6 +110,13 @@ fn memory_with_bootloader_write_test() {
 fn memory_test_parallel_accesses() {
     let f = "std/memory_test_parallel_accesses.asm";
     regular_test(f, &[]);
+}
+
+#[test]
+#[ignore = "Too slow"]
+fn memory16_test() {
+    let f = "std/memory16_test.asm";
+    test_plonky3_with_backend_variant::<BabyBearField>(f, vec![], BackendVariant::Composite);
 }
 
 #[test]
