+++ std/machines/mod.asm
@@ -1,4 +1,3 @@
-mod add_sub16;
 mod arith;
 mod arith16;
 mod binary;
@@ -6,6 +5,7 @@ mod binary_bb;
 mod range;
 mod hash;
 mod memory;
+mod memory16;
 mod memory_with_bootloader_write;
 mod shift;
 mod shift16;
